changes to script

+++ clean_log_files.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 
 #Cleaning log files
-rm $OLDATA/consensus.log-*.gz
+
+if [ -f '$OLDATA/consensus.log-*.gz' ]
+then
+  echo "File does not exist. Skipping..."
+else
+  rm $OLDATA/consensus.log-*.gz
+fi
